Extract DataTableFilter
